Added layout page and fixed svg
